LLM7_OllamaRAG2/requirements.txt
llama-index-llms-ollama
llama-index
llama-index-embeddings-ollama
streamlit
